[Feat] 날짜 변환 함수 구현, 마일스톤 조회 api 연동 날짜 형식 변환해주는 함수 utils로 분리 List 컴포넌트에서 api 호출 -> Milestone 컴포넌트로 넘겨줌 Header 컴포넌트에 open, close 이슈 개수 반영
