+++ client/src/components/milestones/Header/index.js
@@ -5,19 +5,20 @@ import S from './style';
 
 function Header() {
   const history = useHistory();
+
   return (
     <S.MilestoneHeader>
       <S.LabelMilestone>
-        <S.LabelsButton onClick={()=>history.push('/label')}>
+        <S.LabelsButton onClick={() => history.push('/label')}>
           <TagIcon size={14} />
           <S.ButtonText>Labels</S.ButtonText>
         </S.LabelsButton>
-        <S.MilestonesButton onClick={()=>history.push('/milestone')}>
+        <S.MilestonesButton onClick={() => history.push('/milestone')}>
           <MilestoneIcon />
           <S.ButtonText>Milestones</S.ButtonText>
         </S.MilestonesButton>
       </S.LabelMilestone>
-      <S.NewMilestoneButton>New Milestone</S.NewMilestoneButton>
+      <S.NewMilestoneButton onClick={() => history.push('/milestone/new')}>New Milestone</S.NewMilestoneButton>
     </S.MilestoneHeader>
   );
 }
