+++ client/src/utils/date.js
@@ -1,0 +1,5 @@
+const getDate = (date, options) => {
+  return new Date(date).toLocaleDateString('en-US', options);
+};
+
+export default { getDate };
